Formatted and linted Session 7 scratch diagnostics
* cargo fmt auto-fixes in batch.rs and pipeline.rs — Session 7's
  diagnostic code was never fmt'd before commit 2577e61.
* Fixed needless_range_loop clippy error in batch.rs grid dump.
* Fixed f32 -> f64 test type annotation in lib.rs (preprocessed
  buffer was flipped to f64 in 2577e61 but the test was missed).

+++ crates/osprey-scoring/src/lib.rs
@@ -4624,13 +4624,13 @@ mod tests {
 
         // Preprocessed XCorr: manually compute what xcorr does
         let preprocessed = scorer.preprocess_spectrum_for_xcorr(&spectrum);
-        let xcorr_raw: f32 = entry
+        let xcorr_raw: f64 = entry
             .fragments
             .iter()
             .filter_map(|frag| scorer.bin_config().mz_to_bin(frag.mz))
             .map(|bin| preprocessed[bin])
             .sum();
-        let xcorr_preprocessed = (xcorr_raw * 0.005) as f64;
+        let xcorr_preprocessed = xcorr_raw * 0.005;
 
         assert!(
             (direct_score.xcorr - xcorr_preprocessed).abs() < 1e-6,
